+++ packages/categorize/configure/src/design/choices/__tests__/index.test.jsx
@@ -2,15 +2,16 @@ import { shallow } from 'enzyme';
 import React from 'react';
 
 import { Choices } from '../index';
+import defaults from '../../../defaults';
 
 describe('choices', () => {
-  let onChange;
+  let onModelChanged = jest.fn();
+  let model = defaults.model;
   let onConfigChange;
   let onAdd;
   let onDelete;
 
   beforeEach(() => {
-    onChange = jest.fn();
     onConfigChange = jest.fn();
     onAdd = jest.fn();
     onDelete = jest.fn();
@@ -18,15 +19,10 @@ describe('choices', () => {
 
   const wrapper = extras => {
     const props = {
-      onChange,
-      onConfigChange,
-      onAdd,
-      onDelete,
+      onModelChanged,
+      model,
       classes: {},
-      choices: [{ id: '1', content: 'content' }],
-      config: {
-        columns: 2
-      },
+      choices: [{ id: '0', content: 'Choice 0' }],
       ...extras
     };
     return shallow(<Choices {...props} />);
@@ -41,41 +37,11 @@ describe('choices', () => {
   });
 
   describe('logic', () => {
-    describe('toggleRemoveAllTiles', () => {
-      it('adds categoryCount as 1 to choices if it was undefined', () => {
-        w = wrapper();
-        w.instance().toggleRemoveAllTiles();
-        expect(onChange).toBeCalledWith([
-          { id: '1', content: 'content', categoryCount: 1 }
-        ]);
-      });
-
-      it('sets categoryCount to 1 in every choice if it was 0', () => {
-        w = wrapper({
-          choices: [{ id: '1', content: 'content', categoryCount: 0 }]
-        });
-        w.instance().toggleRemoveAllTiles();
-        expect(onChange).toBeCalledWith([
-          { id: '1', content: 'content', categoryCount: 1 }
-        ]);
-      });
-
-      it('sets categoryCount to 0 in every choice if it was 1', () => {
-        w = wrapper({
-          choices: [{ id: '1', content: 'content', categoryCount: 1 }]
-        });
-        w.instance().toggleRemoveAllTiles();
-        expect(onChange).toBeCalledWith([
-          { id: '1', content: 'content', categoryCount: 0 }
-        ]);
-      });
-    });
-
     describe('changeChoice', () => {
-      it('calls onChange with updated choice', () => {
+      it('calls onModelChanged with updated choice', () => {
         w = wrapper();
-        w.instance().changeChoice({ id: '1', content: 'update' });
-        expect(onChange).toBeCalledWith([{ id: '1', content: 'update' }]);
+        w.instance().changeChoice({ id: '0', content: 'update' });
+        expect(onModelChanged).toBeCalledWith({ choices: [{ id: '0', content: 'update' }] });
       });
     });
 
@@ -86,7 +52,7 @@ describe('choices', () => {
       });
 
       it('returns true if all choices have count 1', () => {
-        w = wrapper({ choices: [{ id: '1', categoryCount: 1 }] });
+        w = wrapper({ choices: [{ id: '0', categoryCount: 1 }] });
         expect(w.instance().allChoicesHaveCount(1)).toEqual(true);
       });
 
@@ -94,6 +60,23 @@ describe('choices', () => {
         w = wrapper({ choices: [{ id: '0' }, { id: '1', categoryCount: 1 }] });
         expect(w.instance().allChoicesHaveCount(1)).toEqual(false);
       });
+    });
+
+    describe('addChoice', () => {
+      w = wrapper();
+      w.instance().addChoice();
+
+      expect(onModelChanged).toBeCalledWith({ choices: [{ id: '0', content: 'Choice 0' }, { id: '1', content: 'Choice 1' }] });
+    });
+
+    describe('deleteChoice', () => {
+      w = wrapper();
+      w.instance().deleteChoice({ id: '0' });
+
+      expect(onModelChanged).toBeCalledWith(expect.objectContaining({
+        choices: [],
+        correctResponse: []
+      }));
     });
   });
 });