+++ packages/categorize/configure/src/index.js
@@ -11,13 +11,14 @@ import defaults from './defaults';
 
 export default class CategorizeConfigure extends HTMLElement {
   static createDefaultModel = (model = {}) => ({
-    ...defaults,
+    ...defaults.model,
     ...model,
   });
 
   constructor() {
     super();
     this._model = CategorizeConfigure.createDefaultModel();
+    this._configuration = defaults.configuration;
   }
 
   set model(m) {
@@ -25,10 +26,21 @@ export default class CategorizeConfigure extends HTMLElement {
     this.render();
   }
 
-  onChange(m) {
+  set configuration(c) {
+    this._configuration = c;
+    this.render();
+  }
+
+  onModelChanged(m) {
     this._model = m;
 
+    this.render();
     this.dispatchEvent(new ModelUpdatedEvent(this._model, true));
+  }
+
+  onConfigurationChanged(c) {
+    this._configuration = c;
+    this.render();
   }
 
   connectedCallback() {}
@@ -48,7 +60,9 @@ export default class CategorizeConfigure extends HTMLElement {
   render() {
     const el = React.createElement(Main, {
       model: this._model,
-      onChange: this.onChange.bind(this),
+      configuration: this._configuration,
+      onModelChanged: this.onModelChanged.bind(this),
+      onConfigurationChanged: this.onConfigurationChanged.bind(this),
       imageSupport: {
         add: this.insertImage.bind(this),
         delete: this.onDeleteImage.bind(this)