+++ packages/number-line/configure/src/defaults.js
@@ -1,22 +1,31 @@
 export default {
-  correctResponse: [],
-  config: {
-    domain: [-5, 5],
-    initialElements: [],
-    showMinorTicks: true,
-    initialType: 'PF',
-    exhibitOnly: false,
-    availableTypes: {
-      PF: true,
-      PE: false,
-      LFF: false,
-      LEF: false,
-      LFE: false,
-      LEE: false,
-      RFN: false,
-      RFP: false,
-      REN: false,
-      REP: false
+  model: {
+    prompt: 'Question Prompt goes here.',
+    correctResponse: [],
+    graph: {
+      domain: [-5, 5],
+      initialElements: [],
+      showMinorTicks: true,
+      initialType: 'PF',
+      exhibitOnly: false,
+      availableTypes: {
+        PF: true,
+        PE: false,
+        LFF: false,
+        LEF: false,
+        LFE: false,
+        LEE: false,
+        RFN: false,
+        RFP: false,
+        REN: false,
+        REP: false
+      }
+    }
+  },
+  configuration: {
+    prompt: {
+      settings: true,
+      label: 'Prompt'
     }
   }
 };