+++ packages/graph-lines/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.0.0...@pie-element/graph-lines-configure@2.0.1) (2019-05-08)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
 # [2.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@1.3.5...@pie-element/graph-lines-configure@2.0.0) (2019-05-08)