+++ packages/match/configure/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@2.1.0...@pie-element/match-configure@2.1.1) (2019-05-14)
+
+**Note:** Version bump only for package @pie-element/match-configure
+
+
+
+
+
+# [2.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@2.0.3...@pie-element/match-configure@2.1.0) (2019-05-14)
+
+
+### Features
+
+* **match:** added prompt support. ([f5f692d](https://github.com/pie-framework/pie-elements/commit/f5f692d))
+
+
+
+
+
+## [2.0.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@2.0.2...@pie-element/match-configure@2.0.3) (2019-05-10)
+
+**Note:** Version bump only for package @pie-element/match-configure
+
+
+
+
+
+## [2.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@2.0.1...@pie-element/match-configure@2.0.2) (2019-05-10)
+
+
+### Bug Fixes
+
+* **match:** added configuration. ([04ee245](https://github.com/pie-framework/pie-elements/commit/04ee245))
+
+
+
+
+
+## [2.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@2.0.0...@pie-element/match-configure@2.0.1) (2019-05-08)
+
+**Note:** Version bump only for package @pie-element/match-configure
+
+
+
+
+
 # [2.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-configure@1.8.5...@pie-element/match-configure@2.0.0) (2019-05-08)