+++ packages/match/controller/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.1.0...@pie-element/match-controller@2.1.1) (2019-05-14)
+
+**Note:** Version bump only for package @pie-element/match-controller
+
+
+
+
+
+# [2.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.0.1...@pie-element/match-controller@2.1.0) (2019-05-14)
+
+
+### Features
+
+* **match:** added prompt support. ([f5f692d](https://github.com/pie-framework/pie-elements/commit/f5f692d))
+
+
+
+
+
+## [2.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@2.0.0...@pie-element/match-controller@2.0.1) (2019-05-08)
+
+**Note:** Version bump only for package @pie-element/match-controller
+
+
+
+
+
 # [2.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@1.3.1...@pie-element/match-controller@2.0.0) (2019-05-08)