+++ packages/pie-models/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@4.2.0...@pie-element/pie-models@4.3.0) (2019-05-14)
+
+
+### Features
+
+* **math-inline:** added rationale support. ([7a4f9d4](https://github.com/pie-framework/pie-elements/commit/7a4f9d4))
+
+
+
+
+
+# [4.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@4.1.0...@pie-element/pie-models@4.2.0) (2019-05-14)
+
+
+### Features
+
+* **match:** added prompt support. ([f5f692d](https://github.com/pie-framework/pie-elements/commit/f5f692d))
+* **number-line:** added prompt support. ([324824c](https://github.com/pie-framework/pie-elements/commit/324824c))
+
+
+
+
+
+# [4.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@4.0.1...@pie-element/pie-models@4.1.0) (2019-05-13)
+
+
+### Features
+
+* support paired passages ([#216](https://github.com/pie-framework/pie-elements/issues/216)) ([c036a92](https://github.com/pie-framework/pie-elements/commit/c036a92))
+
+
+
+
+
+## [4.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@4.0.0...@pie-element/pie-models@4.0.1) (2019-05-10)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+# [4.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@3.0.1...@pie-element/pie-models@4.0.0) (2019-05-10)
+
+
+### Features
+
+* **categorize:** Updated with new version of [@pie-lib-config-ui](https://github.com/pie-lib-config-ui). Removed `model.config` and instead added new props directly on model: `choicesPerRow`, `categoriesPerRow`, `choicesPosition`, `choicesLabel`, `lockChoiceOrder`, `removeTilesAfterPlacing`. ([5922f85](https://github.com/pie-framework/pie-elements/commit/5922f85))
+
+
+### BREAKING CHANGES
+
+* **categorize:** `model.config` is removed, use `choicesPerRow`, `categoriesPerRow`, `choicesPosition`, `choicesLabel`, `lockChoiceOrder`, `removeTilesAfterPlacing` on model instead.
+
+
+
+
+
+## [3.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@3.0.0...@pie-element/pie-models@3.0.1) (2019-05-08)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
 # [3.0.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@2.0.0...@pie-element/pie-models@3.0.0) (2019-05-08)